blaan/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
pymongo==4.5.0
python-dotenv==1.0.0
newsapi==0.1.1
requests==2.31.0
twilio==8.10.0
python-telegram-bot==20.3
email-validator==2.1.0
APScheduler==3.10.4
gunicorn==21.2.0
bcrypt==4.1.1
pytz==2023.3
